enlarge input type for order n

+++ src/cl.rs
@@ -7,12 +7,12 @@ pub trait Cl<T> {
     ///
     /// println!("Cl_{}({}) = {}", 2, 1.0, 1.0.cl(2));
     /// ```
-    fn cl(&self, n: i32) -> T;
+    fn cl(&self, n: i64) -> T;
 }
 
 
 impl Cl<f64> for f64 {
-    fn cl(&self, n: i32) -> f64 {
+    fn cl(&self, n: i64) -> f64 {
         match n {
             1 => cl1(*self),
             _ => 0.0
